+++ packages/libp2p-daemon-server/CHANGELOG.md
@@ -1,3 +1,68 @@
+## [@libp2p/daemon-server-v3.0.0](https://github.com/libp2p/js-libp2p-daemon/compare/@libp2p/daemon-server-v2.0.4...@libp2p/daemon-server-v3.0.0) (2022-09-09)
+
+
+### ⚠ BREAKING CHANGES
+
+* the stream type returned by `client.openStream` has changed
+
+### Bug Fixes
+
+* allow opening remote streams ([#126](https://github.com/libp2p/js-libp2p-daemon/issues/126)) ([361cc57](https://github.com/libp2p/js-libp2p-daemon/commit/361cc5750de505ab0381ae43609c67d5d4f659a7))
+
+
+### Dependencies
+
+* update sibling dependencies ([c3ebd58](https://github.com/libp2p/js-libp2p-daemon/commit/c3ebd588abc36ef45667e8e4e4c0e220303b7510))
+
+## [@libp2p/daemon-server-v2.0.4](https://github.com/libp2p/js-libp2p-daemon/compare/@libp2p/daemon-server-v2.0.3...@libp2p/daemon-server-v2.0.4) (2022-08-10)
+
+
+### Bug Fixes
+
+* update all deps ([#124](https://github.com/libp2p/js-libp2p-daemon/issues/124)) ([5e46e1e](https://github.com/libp2p/js-libp2p-daemon/commit/5e46e1e26c23428046a6007ab158420d3d830145))
+
+
+### Documentation
+
+* readme update ([f569ffc](https://github.com/libp2p/js-libp2p-daemon/commit/f569ffc5c3956248e685d99904408fd3f4d868f4))
+
+## [@libp2p/daemon-server-v2.0.3](https://github.com/libp2p/js-libp2p-daemon/compare/@libp2p/daemon-server-v2.0.2...@libp2p/daemon-server-v2.0.3) (2022-07-31)
+
+
+### Trivial Changes
+
+* update project config ([#111](https://github.com/libp2p/js-libp2p-daemon/issues/111)) ([345e663](https://github.com/libp2p/js-libp2p-daemon/commit/345e663e34278e780fc2f3a6b595294f925c4521))
+
+
+### Dependencies
+
+* update uint8arraylist and protons deps ([#115](https://github.com/libp2p/js-libp2p-daemon/issues/115)) ([34a8334](https://github.com/libp2p/js-libp2p-daemon/commit/34a83340ba855a9c08319ae1cd735dfa8b71c248))
+
+## [@libp2p/daemon-server-v2.0.2](https://github.com/libp2p/js-libp2p-daemon/compare/@libp2p/daemon-server-v2.0.1...@libp2p/daemon-server-v2.0.2) (2022-06-17)
+
+
+### Trivial Changes
+
+* update deps ([#105](https://github.com/libp2p/js-libp2p-daemon/issues/105)) ([0bdab0e](https://github.com/libp2p/js-libp2p-daemon/commit/0bdab0ee254e32d6dca0e5fe239d4ef16db41b87))
+
+## [@libp2p/daemon-server-v2.0.1](https://github.com/libp2p/js-libp2p-daemon/compare/@libp2p/daemon-server-v2.0.0...@libp2p/daemon-server-v2.0.1) (2022-06-15)
+
+
+### Trivial Changes
+
+* update deps ([#103](https://github.com/libp2p/js-libp2p-daemon/issues/103)) ([2bfaa37](https://github.com/libp2p/js-libp2p-daemon/commit/2bfaa37e2f056dcd5de5a3882b77f52553c595d4))
+
+## [@libp2p/daemon-server-v2.0.0](https://github.com/libp2p/js-libp2p-daemon/compare/@libp2p/daemon-server-v1.0.5...@libp2p/daemon-server-v2.0.0) (2022-06-15)
+
+
+### ⚠ BREAKING CHANGES
+
+* uses new single-issue libp2p interface modules
+
+### Features
+
+* update to latest libp2p interfaces ([#102](https://github.com/libp2p/js-libp2p-daemon/issues/102)) ([f5e9121](https://github.com/libp2p/js-libp2p-daemon/commit/f5e91210654ab3c411e316c1c657356c037a0f6a))
+
 ## [@libp2p/daemon-server-v1.0.5](https://github.com/libp2p/js-libp2p-daemon/compare/@libp2p/daemon-server-v1.0.4...@libp2p/daemon-server-v1.0.5) (2022-05-25)